+++ src/alignment/phmm/mod.rs
@@ -139,7 +139,8 @@ impl<T: Float, const S: usize> Default for LayerParams<T, S> {
 }
 
 // TODO: Fix doc comments (add local)
-/// The core profile hidden Markov model (pHMM) used by [`GlobalPhmm`] and other PHMMs.
+/// The core profile hidden Markov model (pHMM) used by [`GlobalPhmm`],
+/// [`LocalPhmm`] and other PHMMs.
 #[derive(Clone, Eq, PartialEq, Debug)]
 pub struct CorePhmm<T, const S: usize>(
     /// The parameters for the pHMM.
@@ -183,7 +184,8 @@ impl<T, const S: usize> CorePhmm<T, S> {
     }
 }
 
-/// An implementation of a global profile hidden Markov model (pHMM).
+/// An implementation of a profile hidden Markov model (pHMM) for global
+/// alignment (aligning a full sequence to a full model).
 #[derive(Clone, Eq, PartialEq, Debug)]
 pub struct GlobalPhmm<T, const S: usize> {
     /// The mapping used when processing the bases. This will vary depending on
@@ -191,4 +193,58 @@ pub struct GlobalPhmm<T, const S: usize> {
     pub mapping: &'static ByteIndexMap<S>,
     /// The model parameters
     pub core:    CorePhmm<T, S>,
+}
+
+/// An implementation of a profile hidden Markov model (pHMM) for local
+/// alignment (aligning a subsequence to a submodel).
+///
+/// This is created from a [`GlobalPhmm`] using [`into_local_phmm`]. Two
+/// [`LocalModule`] modules are added to either end of the [`GlobalPhmm`], which
+/// can match arbitrarily many bases at the beginning or end of the sequence,
+/// and can skip arbitrarily many states at the beginning or end of the pHMM.
+///
+/// [`into_local_phmm`]: GlobalPhmm::into_local_phmm
+#[derive(Debug)]
+pub struct LocalPhmm<T, const S: usize> {
+    /// The mapping used when processing the bases. This will vary depending on
+    /// the alphabet used.
+    pub mapping: &'static ByteIndexMap<S>,
+    /// The core model containing the parameters.
+    pub core:    CorePhmm<T, S>,
+    /// The module for handling any bases before the core model
+    pub begin:   LocalModule<T, S>,
+    /// The module for handling any bases after the core model
+    pub end:     LocalModule<T, S>,
+}
+
+/// Options for how to construct a [`LocalPhmm`] from a [`GlobalPhmm`]
+#[non_exhaustive]
+pub enum LocalConfig<T, const S: usize> {
+    /// Does not penalize any transitions in the [`LocalModule`], instead solely
+    /// penalizing the emissions for any insertions
+    NoPenalty { background_emission: EmissionParams<T, S> },
+    /// Use a custom [`LocalModule`] for the begin and end.
+    Custom {
+        begin: LocalModule<T, S>,
+        end:   LocalModule<T, S>,
+    },
+}
+
+impl<T: Float, const S: usize> GlobalPhmm<T, S> {
+    #[must_use]
+    pub fn into_local_phmm(self, config: LocalConfig<T, S>) -> LocalPhmm<T, S> {
+        let (begin, end) = match config {
+            LocalConfig::NoPenalty { background_emission } => (
+                LocalModule::no_penalty(&self.core, background_emission.clone()),
+                LocalModule::no_penalty(&self.core, background_emission),
+            ),
+            LocalConfig::Custom { begin, end } => (begin, end),
+        };
+        LocalPhmm {
+            mapping: self.mapping,
+            core: self.core,
+            begin,
+            end,
+        }
+    }
 }